Scrapper/Scrapping/requirements.txt
beautifulsoup4==4.9.2
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.9.27
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
xlrd==1.2.0
